==> notebooks/0a.split_raw_data.ipynb <==
### Split raw [train](https://www.kaggle.com/competitions/amex-default-prediction/data?select=train_data.csv) & [test](https://www.kaggle.com/competitions/amex-default-prediction/data?select=test_data.csv) data into 4 different parquet files
- Spend & Payment
- Risk
- Balance
- Delinquency
#### All the parquet will contains the common columns **customer_ID** and **S_2** (statement_date)

import dask.dataframe as dd
import gc
import os
import pandas as pd
pd.options.display.float_format = "{:,.4f}".format
import sys
import time

from pathlib import Path
rootpath = Path.cwd().parent
sys.path.append(os.path.join(rootpath))

from utils.constants import *
from utils.eda_helpers import get_cols
from utils.extraction_helpers import read_file

%load_ext autoreload
%autoreload

START = time.time()

### Split Train data into 4 different parquet files

%%time
train = dd.read_csv(f"{RAW_DATA_PATH}/train_raw/train_data.csv")

%%time
(
    train.loc[:, ["customer_ID"] + get_cols(train, ["S_", "P_"])].compute()
    .to_parquet(f"{RAW_DATA_PATH}/train_raw/spend_payment_only.parquet")
)

%%time
(
    train.loc[:, ["customer_ID"] + get_cols(train, "R_")].compute()
    .to_parquet(f"{RAW_DATA_PATH}/train_raw/risk_only.parquet")
)

gc.collect()

%%time
(
    train.loc[:, ["customer_ID"] + get_cols(train, "B_")].compute()
    .to_parquet(f"{RAW_DATA_PATH}/train_raw/balance_only.parquet")
)

%%time
(
    train.loc[:, ["customer_ID"] + get_cols(train, "D_")].compute()
    .to_parquet(f"{RAW_DATA_PATH}/train_raw/delinquency_only.parquet")
)

gc.collect()

del train

### Split Test data into 4 different parquet files

%%time
test = dd.read_csv(f"{RAW_DATA_PATH}/test_raw/test_data.csv")

%%time
(
    test.loc[:, ["customer_ID"] + get_cols(test, ["S_", "P_"])].compute()
    .to_parquet(f"{RAW_DATA_PATH}/test_raw/spend_payment_only.parquet")
)

%%time
(
    test.loc[:, ["customer_ID"] + get_cols(test, "R_")].compute()
    .to_parquet(f"{RAW_DATA_PATH}/test_raw/risk_only.parquet")
)

gc.collect()

%%time
(
    test.loc[:, ["customer_ID"] + get_cols(test, "B_")].compute()
    .to_parquet(f"{RAW_DATA_PATH}/test_raw/balance_only.parquet")
)

%%time
(
    test.loc[:, ["customer_ID"] + get_cols(test, "D_")].compute()
    .to_parquet(f"{RAW_DATA_PATH}/test_raw/delinquency_only.parquet")
)

gc.collect()

del test

END = time.time()

print(f"{END - START:.2f} seconds elapsed")
