==> notebooks/1f.combine_data.ipynb <==
import gc
import math
import matplotlib.pyplot as plt
import numpy as np
import os
import pandas as pd
pd.options.display.float_format = "{:,.4f}".format
import random
import sys
import time
import warnings
warnings.simplefilter("ignore")
warnings.filterwarnings("ignore")
from pandarallel import pandarallel
from tqdm import tqdm

from pathlib import Path
rootpath = Path.cwd().parent
sys.path.append(os.path.join(rootpath))

from utils.constants import *
from utils.eda_helpers import (
    plot_missing_proportion_barchart, get_cols, plot_scatterplot, plot_target_check, 
    plot_int_feature_distribution, plot_train_test_distribution, check_overlap_missing,
    insert_row_number, plot_sampled_time_series
)
from utils.eval_helpers import (
    plot_roc_curves, plot_feature_importance, 
    amex_metric, get_final_metric_df, amex_metric_np, lgb_amex_metric
)
from utils.extraction_helpers import read_file
from utils.feature_group import (
    CATEGORY_COLUMNS, CONTINUOUS_COLUMNS, BINARY_COLUMNS,
    MEAN_FEATURES, MIN_FEATURES, MAX_FEATURES, LAST_FEATURES, NON_FEATURE_COLUMNS
)
from utils.preprocess_helpers import *

%load_ext autoreload
%autoreload

gc.collect()

START = time.time()

### Combine Train Data

%%time
train_sp = read_file(f"{INTERIM_TRAIN_PARQUET_PATH}/spend_payment.parquet", replace_negative127=False)
train_r = read_file(f"{INTERIM_TRAIN_PARQUET_PATH}/risk.parquet", replace_negative127=False)
train_b = read_file(f"{INTERIM_TRAIN_PARQUET_PATH}/balance.parquet", replace_negative127=False)
train_d1 = read_file(f"{INTERIM_TRAIN_PARQUET_PATH}/delinquency_part1.parquet", replace_negative127=False)
train_d2 = read_file(f"{INTERIM_TRAIN_PARQUET_PATH}/delinquency_part2.parquet", replace_negative127=False)

%%time
train = pd.concat([
    train_sp.set_index(["customer_ID", "S_2", "target"]),
    train_r.set_index(["customer_ID", "S_2", "target"]),
    train_b.set_index(["customer_ID", "S_2", "target"]),
    train_d1.set_index(["customer_ID", "S_2", "target"]),
    train_d2.set_index(["customer_ID", "S_2", "target"])
], axis=1).reset_index()

train.shape

train = train.replace(-127, np.nan)

# train.to_parquet(f"{INTERIM_TRAIN_PARQUET_PATH}/all_variables.parquet")

# train.head()

# del train, train_sp, train_r, train_b, train_d1, train_d2

### Combine Test Data

%%time
# test_master = read_file(f"{INTERIM_TEST_PARQUET_PATH}/test_psrbd.parquet", replace_negative127=True)
test_sp = read_file(f"{INTERIM_TEST_PARQUET_PATH}/spend_payment.parquet", replace_negative127=False)
test_r = read_file(f"{INTERIM_TEST_PARQUET_PATH}/risk.parquet", replace_negative127=False)
test_b = read_file(f"{INTERIM_TEST_PARQUET_PATH}/balance.parquet", replace_negative127=False)
test_d1 = read_file(f"{INTERIM_TEST_PARQUET_PATH}/delinquency_part1.parquet", replace_negative127=False)
test_d2 = read_file(f"{INTERIM_TEST_PARQUET_PATH}/delinquency_part2.parquet", replace_negative127=False)

%%time
test = pd.concat([
    test_sp.set_index(["customer_ID", "S_2"]),
    test_r.set_index(["customer_ID", "S_2"]),
    test_b.set_index(["customer_ID", "S_2"]),
    test_d1.set_index(["customer_ID", "S_2"]),
    test_d2.set_index(["customer_ID", "S_2"])
], axis=1).reset_index()

test.shape

test = test.replace(-127, np.nan)

# test.to_parquet(f"{INTERIM_TEST_PARQUET_PATH}/all_variables.parquet")

# test.head()

# del test, test_sp, test_r, test_b, test_d1, test_d2
