==> notebooks/1z.final_clean.ipynb <==
### 1. Clean Parquet data from Kaggle (integer dtypes parquet)
- This notebook aims to further clean the data taken from Kaggle's user, without affecting data quality

import pandas as pd
pd.options.display.float_format = "{:,.4f}".format
import matplotlib.pyplot as plt
import gc
import numpy as np
import os
import sys
import warnings
warnings.simplefilter("ignore")
warnings.filterwarnings("ignore")

from pathlib import Path
rootpath = Path.cwd().parent
sys.path.append(os.path.join(rootpath))

from utils.eval_helpers import (
    plot_roc_curves, plot_feature_importance, 
    amex_metric, get_final_metric_df, amex_metric_np, lgb_amex_metric
)
from utils.eda_helpers import (
    plot_missing_proportion_barchart, get_cols, plot_scatterplot, plot_target_check, 
    plot_int_feature_distribution, plot_train_test_distribution, check_overlap_missing,
    insert_row_number
)
from utils.extraction_helpers import read_file
from utils.feature_group import (
    CATEGORY_COLUMNS, CONTINUOUS_COLUMNS, BINARY_COLUMNS, 
    MEAN_FEATURES, MIN_FEATURES, MAX_FEATURES, LAST_FEATURES, NON_FEATURE_COLUMNS
)
from utils.constants import (
    PROCESSED_DATA_PATH
)
from utils.feature_engineering_helpers import integerize

RAW_DATA_PATH = "../raw_data"
PROCESSED_DATA_PATH = "../processed_data"
SUBMISSION_DATA_PATH = "../submissions"
EVALUATION_DATA_PATH = "../evaluation_data"
MODELS_PATH = "../models"

RAW_TRAIN_PARQUET_PATH = os.path.join(RAW_DATA_PATH, "train_parquet")
RAW_TRAIN_PICKLE_PATH = os.path.join(RAW_DATA_PATH, "train_pickle")
RAW_TEST_PARQUET_PATH = os.path.join(RAW_DATA_PATH, "test_parquet")
RAW_TEST_PICKLE_PATH = os.path.join(RAW_DATA_PATH, "test_pickle")

%load_ext autoreload
%autoreload

### Read Data

%%time
train = read_file(f"{RAW_TRAIN_PARQUET_PATH}/train_bdpsr.parquet", replace_negative_one=True)
test = read_file(f"{RAW_TEST_PARQUET_PATH}/test_bdpsr.parquet", replace_negative_one=True)

labels = pd.read_csv(f"{RAW_DATA_PATH}/train_labels.csv")

%%time
if "target" not in train.columns:
    train = train.merge(labels, on="customer_ID", how="left")

%%time
train.loc[:, CATEGORY_COLUMNS] = train.loc[:, CATEGORY_COLUMNS].astype("category")
test.loc[:, CATEGORY_COLUMNS] = test.loc[:, CATEGORY_COLUMNS].astype("category")

%%time
insert_row_number(train)
insert_row_number(test)

train = train.drop(columns="temp", errors="ignore")
test = test.drop(columns="temp", errors="ignore")

train.shape, test.shape

%%time
train.to_pickle(f"{RAW_TRAIN_PICKLE_PATH}/raw_train_bdpsr.pkl")
test.to_pickle(f"{RAW_TEST_PICKLE_PATH}/raw_test_bdpsr.pkl")
