==> notebooks/3a.feature_engineering_lgbm.ipynb <==
import datetime
import gc
import joblib
import matplotlib.pyplot as plt
import pandas as pd
import numpy as np
import os
import scipy.stats
import seaborn as sns
import sys
sys.path.append("../")
import time
import warnings
warnings.simplefilter("ignore")
from itertools import repeat
from sklearn.preprocessing import StandardScaler
from tqdm import tqdm

from utils.common import (
    sigmoid, pad_column_name
)
from utils.constants import (
    RAW_DATA_PATH, PROCESSED_DATA_PATH
)
from utils.eval_helpers import (
    plot_roc_curves, plot_feature_importance, 
    amex_metric, get_final_metric_df, amex_metric_np, lgb_amex_metric
)
from utils.eda_helpers import (
    plot_missing_proportion_barchart, plot_heatmap,
    get_cols
)
from utils.extraction_helpers import read_file
from utils.feature_group import (
    CATEGORY_COLUMNS, CONTINUOUS_COLUMNS, NON_FEATURE_COLUMNS,
    MEAN_FEATURES, MIN_FEATURES, MAX_FEATURES, LAST_FEATURES, FIRST_FEATURES,
    RANGE_FEATURES, VELOCITY_FEATURES, SPEED_FEATURES
)
from utils.feature_engineering_helpers import (
    filter_df_for_feature, 
    get_specific_row_df, 
    get_agg_df, get_ma_df
)

RAW_DATA_PATH = "../raw_data"
PROCESSED_DATA_PATH = "../processed_data"
EVALUATION_DATA_PATH = "../evaluation_data"
MODELS_PATH = "../models"
EXP_PATH = "../experiments"

RAW_TRAIN_PICKLE_PATH = os.path.join(RAW_DATA_PATH, "train_pickle")
RAW_TEST_PICKLE_PATH = os.path.join(RAW_DATA_PATH, "test_pickle")

%load_ext autoreload
%autoreload

## Feature Engineering on Train

### Read master data & train labels

%%time
raw_train = read_file(f"{RAW_TRAIN_PICKLE_PATH}/raw_train_data.pkl")
labels = read_file(f"{RAW_DATA_PATH}/train_labels.csv")

# plot_missing_proportion_barchart(raw_train, top_n=115, figsize=(18, 35), title="Missing proportion barchart")

# numeric_columns = raw_train.select_dtypes(np.number).drop(columns=NON_FEATURE_COLUMNS, errors='ignore').columns.tolist()

# raw_train.loc[:, ["customer_ID"] + numeric_columns].groupby(["customer_ID"]).diff()

# %%time
# raw_train["P_200"] = raw_train["P_2"] - raw_train.groupby("customer_ID")["P_2"].transform(lambda x: x.ewm(span=13).mean()).values
# raw_train["P_300"] = raw_train["P_3"] - raw_train.groupby("customer_ID")["P_3"].transform(lambda x: x.ewm(span=13).mean()).values
# raw_train["S_300"] = raw_train["S_3"] - raw_train.groupby("customer_ID")["S_3"].transform(lambda x: x.ewm(span=13).mean()).values
# raw_train["S_1200"] = raw_train["S_12"] - raw_train.groupby("customer_ID")["S_12"].transform(lambda x: x.ewm(span=13).mean()).values

# float_columns = raw_train.select_dtypes(np.float).columns
# low_missing_thr = 0.03
# low_missing_columns = raw_train.columns[raw_train.isnull().sum() < raw_train.shape[0] * low_missing_thr]
# high_std_thr = 0.12
# high_std_float_columns = float_columns[raw_train[float_columns].std() >= high_std_thr]

# column_to_calc_diff = set(float_columns).intersection(set(low_missing_columns), set(high_std_float_columns))
# len(column_to_calc_diff)

raw_train.shape

# %%time
# for column in tqdm(column_to_calc_diff):
#     raw_train[f"{column}_pct_diff"] = raw_train.groupby("customer_ID")[column].pct_change()

### Get simple aggregation values (average, min, max, std)

%%time
train_agg = get_agg_df(raw_train)

train_agg.shape

### Calculate number of statements (data availability for each client)

%%time
train_agg["num_statements"] = (
    raw_train.loc[raw_train["row_number"] == 1][["row_number", "row_number_inv"]].sum(axis=1) - 1
).reset_index(drop=True).values

### Get Last, Second-Last, Third-Last, First

%%time
train_last_etc = get_specific_row_df(raw_train)

# train_last_etc.head(3)

%%time
train_agg = train_last_etc.merge(train_agg, left_index=True, right_index=True, how="inner")
del train_last_etc

### Get Moving Average columns (MA3_R1, MA3_R2, MA3_R3, MA3_R4)

gc.collect()

%%time
train_ma_df = get_ma_df(raw_train)

%%time
train_agg = train_agg.merge(train_ma_df, left_index=True, right_index=True, how="inner")
del train_ma_df

numeric_columns = list(set(raw_train.columns) - set(CATEGORY_COLUMNS) - set(NON_FEATURE_COLUMNS))
all_columns = list(set(numeric_columns).union(set(CATEGORY_COLUMNS)))
del raw_train

train_agg.shape

### Feature Crossing between aggregated features

%%time
for col in tqdm(numeric_columns):
    train_agg[f"{col}_range"] = train_agg[f"{col}_max"] - train_agg[f"{col}_min"]
    train_agg[f"{col}_displacement"] = train_agg[f"{col}_last"] - train_agg[f"{col}_first"]
    train_agg[f"{col}_velocity"] = train_agg[f"{col}_displacement"] / train_agg["num_statements"]
    train_agg[f"{col}_sprint"] = train_agg[f"{col}_last"] - train_agg[f"{col}_second_last"]
    train_agg[f"{col}_previous_sprint"] = train_agg[f"{col}_second_last"] - train_agg[f"{col}_third_last"]
    train_agg[f"{col}_acceleration"] = train_agg[f"{col}_sprint"] / (train_agg[f"{col}_previous_sprint"] * train_agg[f"{col}_std"]).replace(
        [np.inf, -np.inf], np.nan
    )
    train_agg[f"{col}_last_minus_avg"] = train_agg[f"{col}_last"] - train_agg[f"{col}_avg"]
    train_agg[f"{col}_coef_var"] = (train_agg[f"{col}_std"] / train_agg[f"{col}_avg"]).replace([np.inf, -np.inf], np.nan)
    train_agg[f"{col}_ma3_r1_r2"] = train_agg[f"{col}_ma3_r1"] / train_agg[f"{col}_ma3_r2"]
    train_agg[f"{col}_ma3_r1_r3"] = train_agg[f"{col}_ma3_r1"] / train_agg[f"{col}_ma3_r3"]
    train_agg[f"{col}_ma3_r1_r4"] = train_agg[f"{col}_ma3_r1"] / train_agg[f"{col}_ma3_r4"]
    train_agg[f"{col}_general_trend"] = 100 * (train_agg[f"{col}_ma3_r1"] - train_agg[f"{col}_ma3_r4"]) / train_agg["num_statements"]
    gc.collect()

train_agg.shape, labels.shape

train_agg["target"] = labels["target"].values


train_agg = train_agg.drop(columns=["num_statements"], errors="ignore")

if "customer_ID" not in train_agg.columns:
    train_agg = train_agg.reset_index().rename(columns={"index": "customer_ID"})

sys.getsizeof(train_agg) / 1e9

train_agg.shape

train_agg.to_pickle(f"{PROCESSED_DATA_PATH}/train_agg_complete2.pkl")

### Feature Engineering on Test

%%time
raw_test = read_file(f"{RAW_TEST_PICKLE_PATH}/raw_test_data.pkl")

%%time
test_agg = get_agg_df(raw_test)

%%time
test_agg["num_statements"] = (
    raw_test.loc[raw_test["row_number"] == 1][["row_number", "row_number_inv"]].sum(axis=1) - 1
).reset_index(drop=True).values

test_agg.head(3)

test_agg.shape

%%time
test_last_etc = get_specific_row_df(raw_test)

test_last_etc.head(3)

%%time
test_agg = test_last_etc.merge(test_agg, left_index=True, right_index=True, how="inner")
del test_last_etc

%%time
test_ma_df = get_ma_df(raw_test)

%%time
test_agg = test_agg.merge(test_ma_df, left_index=True, right_index=True, how="inner")
del test_ma_df

numeric_columns = list(set(raw_test.columns) - set(CATEGORY_COLUMNS) - set(NON_FEATURE_COLUMNS))
all_columns = list(set(numeric_columns).union(set(CATEGORY_COLUMNS)))
del raw_test

test_agg.shape

%%time
for col in tqdm(numeric_columns):
    test_agg[f"{col}_range"] = test_agg[f"{col}_max"] - test_agg[f"{col}_min"]
    test_agg[f"{col}_displacement"] = test_agg[f"{col}_last"] - test_agg[f"{col}_first"]
    test_agg[f"{col}_velocity"] = test_agg[f"{col}_displacement"] / np.log(test_agg["num_statements"])
    test_agg[f"{col}_sprint"] = test_agg[f"{col}_last"] - test_agg[f"{col}_second_last"]
    test_agg[f"{col}_previous_sprint"] = test_agg[f"{col}_second_last"] - test_agg[f"{col}_third_last"]
    test_agg[f"{col}_acceleration"] = test_agg[f"{col}_sprint"] / (test_agg[f"{col}_previous_sprint"] * test_agg[f"{col}_std"]).replace(
        [np.inf, -np.inf], np.nan
    )
    test_agg[f"{col}_last_minus_avg"] = test_agg[f"{col}_last"] - test_agg[f"{col}_avg"]
    test_agg[f"{col}_coef_var"] = (test_agg[f"{col}_std"] / test_agg[f"{col}_avg"]).replace([np.inf, -np.inf], np.nan)
    test_agg[f"{col}_ma3_r1_r2"] = test_agg[f"{col}_ma3_r1"] / test_agg[f"{col}_ma3_r2"]
    test_agg[f"{col}_ma3_r1_r3"] = test_agg[f"{col}_ma3_r1"] / test_agg[f"{col}_ma3_r3"]
    test_agg[f"{col}_ma3_r1_r4"] = test_agg[f"{col}_ma3_r1"] / test_agg[f"{col}_ma3_r4"]
    test_agg[f"{col}_general_trend"] = 100 * (test_agg[f"{col}_ma3_r1"] - test_agg[f"{col}_ma3_r4"]) / test_agg["num_statements"]
    gc.collect()

gc.collect()

%%time
test_agg = test_agg.reset_index().rename(columns={"index": "customer_ID"})
test_agg = test_agg.drop(columns=NON_FEATURE_COLUMNS, errors='ignore')

test_agg.shape

test_agg.dtypes.value_counts()

%%time
test_agg.to_pickle(f"{PROCESSED_DATA_PATH}/test_agg_complete2.pkl")
