==> notebooks/4c.train_catboost.ipynb <==
import catboost as cb
import gc
import joblib
import json
import matplotlib.pyplot as plt
import numpy as np
import optuna
import pandas as pd
import seaborn as sns
import sys
sys.path.append("../")
import warnings
warnings.simplefilter("ignore")
from catboost import CatBoostClassifier, Pool
from sklearn.metrics import roc_auc_score, fbeta_score, make_scorer
from sklearn.model_selection import train_test_split, RandomizedSearchCV, StratifiedKFold
from sklearn.preprocessing import StandardScaler
from tqdm import tqdm

from utils.common import (
    sigmoid, pad_column_name
)
from utils.constants import *
from utils.eval_helpers import (
    plot_roc_curves, plot_feature_importance, 
    amex_metric, get_final_metric_df, amex_metric_np, lgb_amex_metric
)
from utils.eda_helpers import (
    plot_missing_proportion_barchart, 
    get_cols
)
from utils.extraction_helpers import read_file
from utils.feature_group import (
    CATEGORY_COLUMNS, CONTINUOUS_COLUMNS, NON_FEATURE_COLUMNS
)
from utils.feature_engineering_helpers import feature_gen_pipeline

from matplotlib.ticker import MaxNLocator
from matplotlib.colors import ListedColormap
from cycler import cycler
from IPython.display import display
from colorama import Fore, Back, Style
plt.rcParams['axes.facecolor'] = '#0057b8' # blue
plt.rcParams['axes.prop_cycle'] = cycler(color=['#ffd700'] +
                                         plt.rcParams['axes.prop_cycle'].by_key()['color'][1:])
plt.rcParams['text.color'] = 'w'

%load_ext autoreload
%autoreload

### Read Data

train_agg = read_file(f"{PROCESSED_DATA_PATH}/v2/train_agg.pkl")


labels = read_file(f"{RAW_DATA_PATH}/train_labels.csv")

# %%time
# train = pd.concat([
#     spend_payment_df, 
#     risk_df.loc[:, get_cols(risk_df, "R_")]
# ], axis=1)

# train.loc[:, get_cols(train, "R_")] = train.loc[:, get_cols(train, "R_")].fillna(-127)

#### Generate Feature

# %%time
# train_agg, keep_column = feature_gen_pipeline(train)

# train_agg.head(3)

# len(features_to_drop)
# train_agg = train_agg.drop(columns=list(features_to_drop), errors="ignore")

target = labels["target"].values

%%time
train_agg = train_agg.drop(columns=NON_FEATURE_COLUMNS + ["target"], errors="ignore")
gc.collect()

# feature_imp = dict(zip(model.feature_names_, model.feature_importances_))

noob_features = [k for k, v in feature_imp.items() if v < 0.005]

len(noob_features)

train_agg = train_agg.drop(noob_features, errors="ignore")

### Tune Catboost

def objective(trial):
    kf = StratifiedKFold(n_splits=5, shuffle=True, random_state=1020)

    param = {
        "objective": "Logloss",
        "colsample_bylevel": trial.suggest_float("colsample_bylevel", 0.03, 0.1, log=True),
        "depth": trial.suggest_int("depth", 8, 15),
        "random_strength": trial.suggest_float("random_strength", 1.1, 1.5),
        "class_weights": trial.suggest_categorical("class_weights", [[1, 1.25], [1, 1.5], [1, 2]]),
        "n_estimators": trial.suggest_categorical("n_estimators", [500, 750, 1000]),
        "l2_leaf_reg": trial.suggest_float("l2_leaf_reg", 2, 8, log=True),
        "min_data_in_leaf": trial.suggest_categorical("min_data_in_leaf", [1024, 2048, 3072]),
        "learning_rate": trial.suggest_categorical("learning_rate", [0.03, 0.05, 0.08]),
        "max_bin": trial.suggest_categorical("max_bin", [63, 127, 255]),
        "used_ram_limit": "16gb",
        "bootstrap_type": "MVS",
        # "boosting_type": trial.suggest_categorical("boosting_type", ["Ordered", "Plain"]),
        "grow_policy": "Depthwise"
    }

    if param["bootstrap_type"] == "Bernoulli":
        param["subsample"] = trial.suggest_float("subsample", 0.6, 0.85)
    
    print(param)
    train_score_list, val_score_list = [], []
    for fold, (idx_tr, idx_va) in zip(range(1, 5+1), kf.split(train_agg, target)):
        fold = str(fold)
        X_train, y_train = train_agg.iloc[idx_tr], target[idx_tr]
        X_val, y_val = train_agg.iloc[idx_va], target[idx_va]
        with warnings.catch_warnings():
            warnings.filterwarnings('ignore', category=UserWarning)
            model = cb.CatBoostClassifier(**param)
            model.fit(
                X_train, 
                y_train, 
                eval_set=[(X_val, y_val)], 
                verbose=0, 
                cat_features=cat_features,
                early_stopping_rounds=100
            )
        y_train_pred = model.predict(X_train, prediction_type='RawFormulaVal')
        train_score, train_g, train_t4 = amex_metric(y_train, y_train_pred)
        train_data, X_train, y_train = None, None, None
        y_val_pred = model.predict(X_val, prediction_type='RawFormulaVal')
        val_score, val_g, val_t4 = amex_metric(y_val, y_val_pred)
        valid_data, X_val, y_val = None, None, None
        train_score_list.append(train_score)
        val_score_list.append(val_score)
        # if val_score > best_scores_json["validation"][fold]:
        #     best_scores_json["train"][fold] = train_score
        #     best_scores_json["validation"][fold] = val_score
        #     with open(f'{CURRENT_EXP_PATH}/best_scores.json', "w") as outfile:
        #         json.dump(best_scores_json, outfile)
        #     joblib.dump(model, f'{CURRENT_EXP_PATH}/models/model{fold}.pkl')
        # elif np.mean(train_score_list) >= np.mean(list(best_scores_json["train"].values())) + 0.02:
        #     print(f"Train score too high (overfitting), start a new trial")
        #     return np.mean(val_score_list)
        joblib.dump(model, f'{EXP_PATH}/catboost_exp/models/model{fold}_seed1020_{val_score:.5f}.pkl')
        print(f"{Fore.BLUE}{Style.BRIGHT}Fold {fold} | Train Score = {train_score:.5f} ({train_g:.4f}, {train_t4:.4f})")
        print(f"{Fore.GREEN}{Style.BRIGHT}Fold {fold} | Val Score = {val_score:.5f} ({val_g:.4f}, {val_t4:.4f}){Style.RESET_ALL}")
        print(f"Clear cache {gc.collect()}")
        
    return np.mean(val_score_list)

# study = joblib.load(f"{CURRENT_EXP_PATH}/optuna_study.pkl")

study = optuna.create_study(direction="maximize")

study.optimize(objective, n_trials=20)

joblib.dump(study, f"{EXP_PATH}/catboost_exp/catboost_optuna_study.pkl")

study.optimize(objective, n_trials=20)

df = study.trials_dataframe()

df["params_class_weights"] = df["params_class_weights"].astype(str)

df.columns

df.groupby("params_subsample")["value"].mean()

study.best_trial

study_df = study.trials_dataframe()

# for col in get_cols(study_df, "params"):
#     sns.scatterplot(data=study_df, x=study_df[col], y=study_df["value"])
#     plt.show()

### Train Catboost

param = {
    "objective": "Logloss",
    "colsample_bylevel": 0.05,
    "depth": 6,
    "boosting_type": "Ordered",
    "bootstrap_type": "MVS",
    "n_estimators": 1000,
    "l2_leaf_reg": 3,
    "min_data_in_leaf": 1024,
    "learning_rate": 0.05,
    "max_bin": 127,
    "used_ram_limit": "16gb",
}

kf = StratifiedKFold(n_splits=5, shuffle=True, random_state=1020)

train_score_list, val_score_list = [], []
for fold, (idx_tr, idx_va) in zip(range(1, 5+1), kf.split(train_agg, target)):
    fold = str(fold)
    X_train, y_train = train_agg.iloc[idx_tr], target[idx_tr]
    X_val, y_val = train_agg.iloc[idx_va], target[idx_va]
    with warnings.catch_warnings():
        warnings.filterwarnings('ignore', category=UserWarning)
        model = cb.CatBoostClassifier(**param)
        model.fit(
            X_train, 
            y_train, 
            eval_set=[(X_val, y_val)], 
            verbose=100, 
            cat_features=cat_features,
            early_stopping_rounds=100
        )
    y_train_pred = model.predict(X_train, prediction_type='RawFormulaVal')
    train_score, train_g, train_t4 = amex_metric(y_train, y_train_pred)
    train_data, X_train, y_train = None, None, None
    y_val_pred = model.predict(X_val, prediction_type='RawFormulaVal')
    val_score, val_g, val_t4 = amex_metric(y_val, y_val_pred)
    valid_data, X_val, y_val = None, None, None
    train_score_list.append(train_score)
    val_score_list.append(val_score)
    print(f"{Fore.BLUE}{Style.BRIGHT}Fold {fold} | Train Score = {train_score:.5f} ({train_g:.4f}, {train_t4:.4f})")
    print(f"{Fore.GREEN}{Style.BRIGHT}Fold {fold} | Val Score = {val_score:.5f} ({val_g:.4f}, {val_t4:.4f}){Style.RESET_ALL}")
    print(f"Clear cache {gc.collect()}")

# experiment_dict = {}

# experiment_dict["fill_negative_127"] = {"train_scores": train_score_list, "val_scores": val_score_list}

np.mean(experiment_dict["no_impute"]["train_scores"]), np.mean(experiment_dict["no_impute"]["val_scores"])

np.mean(experiment_dict["fill_zero"]["train_scores"]), np.mean(experiment_dict["fill_zero"]["val_scores"])

np.mean(experiment_dict["fill_negative_one"]["train_scores"]), np.mean(experiment_dict["fill_negative_one"]["val_scores"])

np.mean(experiment_dict["fill_negative_127"]["train_scores"]), np.mean(experiment_dict["fill_negative_127"]["val_scores"])
