==> notebooks/5a.lgbm_evaluation.ipynb <==
import catboost
import gc
import joblib
import json
import matplotlib.pyplot as plt
import pandas as pd
import numpy as np
import os
import sys
sys.path.append("../")
import time
import warnings
warnings.simplefilter("ignore")
from lightgbm import LGBMClassifier
from sklearn.model_selection import train_test_split, RandomizedSearchCV, StratifiedKFold
from sklearn.metrics import roc_auc_score, fbeta_score, make_scorer
from tqdm import tqdm

from utils.common import (
    sigmoid, 
    pad_column_name
)
from utils.constants import *
from utils.eval_helpers import (
    plot_roc_curves, plot_feature_importance, 
    amex_metric, get_final_metric_df, amex_metric_np, lgb_amex_metric,
    TreeExperiment
)
from utils.eda_helpers import (
    plot_missing_proportion_barchart, 
    get_cols, plot_heatmap, plot_bar,
)
from utils.extraction_helpers import read_file
from utils.feature_group import (
    CATEGORY_COLUMNS, CONTINUOUS_COLUMNS, NON_FEATURE_COLUMNS
)

%load_ext autoreload
%autoreload

%%time
labels = pd.read_csv(f"{RAW_DATA_PATH}/train_labels.csv")
# if "target" not in train.columns:
#     train = train.merge(labels, on="customer_ID", how="left")

### Read Model

%%time
lgbm_dart = TreeExperiment(
    exp_full_path="../experiments/19.lgbm_dart_923_stratified",
    seed=923, 
    model_path="models"
)

train_agg = read_file(f"{PROCESSED_DATA_PATH}/v2/train_agg.pkl")

# %%time
# df_dict = {}
# for i in tqdm(range(5)):
#     df_dict[i] = read_file(f"{PROCESSED_DATA_PATH}/v10/validation_fold{i}.pkl")

# df_dict[0].loc[:, lgbm_dart.models[0].feature_name()]

# %%time
# t = lgbm_dart.models[0].predict(, raw_score=True)

temp_df = df_dict[0][["customer_ID", "target"]]
temp_df["prediction"] = t

master_df = temp_df.copy()

amex_metric(temp_df["target"], t)

for i in range(1, 5):
    t = lgbm_dart.models[i].predict(df_dict[i].loc[:, lgbm_dart.models[i].feature_name()], raw_score=True)
    temp_df = df_dict[i][["customer_ID", "target"]].copy()
    temp_df["prediction"] = t
    print(f"Fold {i+1}")
    print(amex_metric(temp_df["target"], t))
    master_df = pd.concat([master_df, temp_df], ignore_index=True)

amex_metric(master_df["target"], master_df["prediction"])

master_df.rename(columns={"prediction": "oof_score"}).to_csv(f"{EXP_PATH}/19.lgbm_dart_923_stratified/train_scores_df.csv", index=False)

fdf = lgbm_dart.feature_imp_df.reset_index(drop=True)
fdf.shape

feature_list = fdf.loc[fdf["average_importance"] >= 70, "feature"].tolist()
len(feature_list)

joblib.dump(feature_list, "../evaluation_data/v2_feature_imp.pkl")

# lgbm_dart.models[1].params

# lgbm_dart.models[0].feature_name()

# result = pd.DataFrame(dict(name=lgbm_dart.models[3].feature_name(),
#                            imp=lgbm_dart.models[3].feature_importance()))

# plot_bar(result.nsmallest(300, "imp"), "name", num_col="imp", title=None, horizontal=True, decimal=2, figsize=(15, 15))

# lgbm_dart.feature_imp_summary.to_csv("../experiments/12.lgbm_dart_manual_split_42/feature_importance.csv", index=False)

### Training Inference

%%time
# train_agg = read_file(f"{PROCESSED_DATA_PATH}/v2/train_agg.pkl")
# train_agg = read_file(f"{PROCESSED_DATA_PATH}/v20/train_agg.pkl")

train_agg = pd.DataFrame()
for i in range(5):
    t = read_file(f"{PROCESSED_DATA_PATH}/v19/validation_fold{i}.parquet")
    train_agg = pd.concat([train_agg, t], ignore_index=True)

# lgbm_dart.models[0].predict(train_agg)

# del test_agg

# cat_columns = get_cols(train_agg, CATEGORY_COLUMNS)

# pd.DataFrame(dict(feature=lgbm_dart.models[0].feature_name())).to_csv(f"{PROCESSED_DATA_PATH}/v4/feature_list.csv", index=False)

train_agg.shape

%%time
train_scores_df = lgbm_dart.inference_full(train_agg)

# train_scores_df.to_csv("../experiments/17.lgbm_dart_50/train_scores_df.csv", index=False)

t = read_file("../held_out_predictions/wk_exp19_prediction.csv")

#### If normal stratify K-fold is being use

lgbm_dart.seed

kf = StratifiedKFold(n_splits=5, random_state=lgbm_dart.seed, shuffle=True)
val_indices = [idx_va for idx_tr, idx_va in kf.split(train_agg, labels["target"])]

for i in range(5):
    train_scores_df.loc[val_indices[i], "oof_score"] = train_scores_df.loc[val_indices[i], f"score{i+1}"].values

amex_metric_np(train_scores_df["oof_score"], labels["target"])

# train_scores_df.to_csv("../experiments/16.lgbm_dart_6666_new/train_scores_df.csv", index=False)

#### If manual split is being use

exp12_val_cids = joblib.load(f"./{EXP_PATH}/12.lgbm_dart_manual_split_42/5fold_val_cid.pkl")

for i in range(5):
    train_scores_df.loc[labels["customer_ID"].isin(exp12_val_cids[i]), "oof_score"] = \
        train_scores_df.loc[labels["customer_ID"].isin(exp12_val_cids[i]), f"score{i+1}"].values

# train_scores_df.insert(5, "target", labels["target"])

train_scores_df.to_csv("../experiments/12.lgbm_dart_manual_split_42/train_scores_df.csv", index=False)

# train_scores_df = read_file("../experiments/11.lgbm_dart_round_clip_7788/train_scores_df.csv")

train_scores_df.describe()

# amex_metric_np(train_scores_df["oof_score"], labels["target"])

del train_agg

### Test Inference

%%time
# test_agg = read_file(f"{PROCESSED_DATA_PATH}/v2/test_agg.pkl")
# test_agg = read_file(f"{PROCESSED_DATA_PATH}/v10/held_out.pkl")
test_agg = read_file(f"{PROCESSED_DATA_PATH}/v20/test_agg.pkl")

test_agg.head()

%%time
test_agg = test_agg.reset_index()

test_scores = lgbm_dart.inference_full(test_agg)
# test_scores2 = lgbm_dart.inference_full(test_agg2)

# amex_metric_np(test_scores.mean(axis=1), test_agg["target"])

held_out = test_agg[["customer_ID", "target"]].copy()

held_out["prediction"] = test_scores.mean(axis=1)

# held_out.to_csv("../held_out_predictions/wk_exp20_prediction.csv", index=False)

test_scores.shape

# test_scores["score2"] = test_scores["score2"] * 0.93

test_scores.describe()

test_scores.head()

# test_scores.to_csv(f"../experiments/20.lgbm_dart_6666_stratified/scores_df.csv", index=False)

t = read_file(f"../held_/16.lgbm_dart_6666_new/scores_df.csv").mean(axis=1)

# t = read_file(f"../experiments/12.lgbm_dart_manual_split_42/scores_df.csv").mean(axis=1)

sub = read_file(f"{SUBMISSION_DATA_PATH}/empty.csv")
sub["prediction"] = t
sub.to_csv(f"{SUBMISSION_DATA_PATH}/single_5fold_old_seed6666_dart.csv", index=False)

fi.loc[(fi["last"] - fi["third_last"]) < -15, ["base_feature", "last", "second_last", "third_last"]]

fi2.loc[(fi2["last"] - fi2["first"]) < -5, ["base_feature", "last", "second_last", "third_last", "first"]]

plot_heatmap(fi.corr(), annot=True, figsize=(19, 17), fmt=".2f")

# %%time
# lgbm_gbdt_shuffled = LGBM(
#     exp_full_path="../experiments/lgbm_gbdt_shuffled_exp"
# )
# gbdt_shuffled_master_feature_set = lgbm_gbdt_shuffled.master_feature_set

# master_feature_set = sorted(gbdt_master_feature_set.union(
#     gbdt_shuffled_master_feature_set
# ))
# len(set(master_feature_set))

### Read Train Data (Error Analysis)

labels = read_file(f"{RAW_DATA_PATH}/train_labels.csv")

target = labels["target"]

%%time
raw_train = read_file(f"{RAW_TRAIN_PICKLE_PATH}/raw_train_data.pkl")

%%time
train_agg = train_agg.loc[:, list(gbdt_master_feature_set)]

%%time
train_agg = lgbm_gbdt.inference_fold(train_agg, target)

train_agg.head()

%%time
eval_df = get_final_metric_df(train_agg.iloc[:, :-1], target, train_agg["prediction"])

eval_df.loc[eval_df["is_cutoff"] == 1, "target"].sum() / eval_df["target"].sum()

%%time
false_negative_indices = eval_df.loc[(eval_df["target"] == 1) & (eval_df["is_cutoff"] == 0)].index.tolist()
false_positive_indices = eval_df.loc[(eval_df["target"] == 0) & (eval_df["is_cutoff"] == 1)].index.tolist()

len(false_negative_indices), len(false_positive_indices)

wrong_indices = eval_df.loc[false_negative_indices + false_positive_indices, :].index.tolist()

len(wrong_indices)

%%time
summary_count_df = raw_train.groupby("customer_ID").count()
summary_count_df = summary_count_df.reset_index()

for col in tqdm(summary_count_df.columns[4:]):
    summary_count_df[col] = 1 - (summary_count_df[col] / summary_count_df["row_number"])

wrong_pred_missing_count_df = summary_count_df.loc[wrong_indices]
wrong_pred_missing_count_df = wrong_pred_missing_count_df.sort_index()
correct_pred_missing_count_df = summary_count_df.loc[~summary_count_df.index.isin(wrong_indices)]

correct_mean_df = correct_pred_missing_count_df.iloc[:, 3:].mean()

wrong_mean_df = wrong_pred_missing_count_df.iloc[:, 3:].mean()

((wrong_mean_df / correct_mean_df - 1).sort_values(ascending=False) > 0).sum()

((wrong_mean_df / correct_mean_df - 1).sort_values(ascending=False) < 0).sum()

(wrong_mean_df / (correct_mean_df + wrong_mean_df)).sort_values(ascending=False).to_csv(
    f"{EXP_PATH}/percent_of_missing_from_wrong.csv"
)

wrong_mean_df["D_124"], correct_mean_df["D_124"]

train_, test_ = train_test_split(score_df, test_size=0.2, random_state=1020)

from sklearn.linear_model import LogisticRegression

logreg = LogisticRegression()
logreg.fit(train_.drop(columns="target"), train_["target"])

y_pred = logreg.predict_proba(test_.drop(columns="target"))[:, 1]

y_true = test_["target"].values

amex_metric(y_true, y_pred)

# del train_agg

# lgbm_gbdt.run_validation_check(data=train_agg, labels=labels["target"].values)

lgbm_gbdt.val_metrics

# train_agg = lgbm_dart.run_validation_check(data=train_agg)

# lgbm_dart.val_metrics

### Error Analysis

train_agg.loc[train_agg["target"] == 0, "prediction"].describe()

train_agg.loc[train_agg["target"] == 1, "prediction"].describe()

### High Pred Score Analysis

success_pos_df = train_agg.loc[
    (train_agg["target"] == 1) & (train_agg["is_cutoff"] == 1)
]

fail_neg_df = train_agg.loc[
    (train_agg["target"] == 0) & (train_agg["is_cutoff"] == 1)
]

### Low Pred Score Analysis

success_neg_df = train_agg.loc[
    (train_agg["target"] == 0) & (train_agg["is_cutoff"] == 0)
]

fail_pos_df = train_agg.loc[
    (train_agg["target"] == 1) & (train_agg["is_cutoff"] == 0)
]

success_neg_df["num_statements"].describe()

fail_pos_df["num_statements"].describe()

base_feature = "P_2"

# for col in get_cols(success_neg_df, base_feature):
#     my_dict = {
#         'True Negative': success_neg_df[col].dropna().values, 
#         'False Negative': fail_pos_df[col].dropna().values,
#         'True Positive': success_pos_df[col].dropna().values, 
#     }
#     fig, ax = plt.subplots(figsize=(15, 12))
#     ax.boxplot(my_dict.values())
#     ax.set_xticklabels(my_dict.keys())
#     plt.title(f"{col}")
#     plt.show()

# success_pos_df.to_pickle(f"{EXP_PATH}/lgbm_gbdt_exp3/success_positive.pkl")
# fail_pos_df.to_pickle(f"{EXP_PATH}/lgbm_gbdt_exp3/failed_positive.pkl")
# success_neg_df.to_pickle(f"{EXP_PATH}/lgbm_gbdt_exp3/success_negative.pkl")
# fail_neg_df.to_pickle(f"{EXP_PATH}/lgbm_gbdt_exp3/failed_negative.pkl")

### Read Test Data

%%time
if 'train_agg' in locals():
    del train_agg
# test_agg = read_file(f"{PROCESSED_DATA_PATH}/test_agg_complete2.pkl")
test_agg = read_file(f"{PROCESSED_DATA_PATH}/test_agg.pkl")
gc.collect()

%%time
test_agg = test_agg.loc[:, list(gbdt_master_feature_set)]
test_agg.shape

# lgbm_gbdt.feature_imp_df.iloc[30:60]

%%time
scores_list = []
batch_size = 5000
for j, model in enumerate(lgbm_gbdt.models.values()):
    score_list = []
    print(f"Model {j + 1}")
    for i in tqdm(range(int(test_agg.shape[0] / batch_size) + 1)):
        score_list.append(model.predict(
            test_agg.loc[int(i*batch_size): int((i+1)*batch_size) - 1, model.feature_name()], 
            raw_score=True
        ))
    scores_list.append(np.concatenate(score_list))

with open(f"{lgbm_gbdt.path}/best_scores.json", "r+") as outfile:
    best_scores_json = json.load(outfile)
np.mean(list(best_scores_json["validation"].values()))

# from sklearn.preprocessing import MinMaxScaler

# score_df = pd.DataFrame(np.array(scores_list).T, columns=range(1, 6))
# minmax_scaler = MinMaxScaler()
# score_df = pd.DataFrame(minmax_scaler.fit_transform(score_df), columns=range(1, 6))
# score_df.describe()
# test_agg["prediction"] = np.mean(score_df, axis=1) 

test_agg["prediction"] = np.mean(scores_list, axis=0)

# test_score_df.applymap(sigmoid).mean(axis=1).values # 

test_score_df = pd.DataFrame(np.array(scores_list).T, columns=[f"score{i}" for i in range(1, 6)])

submission = pd.read_csv(f"{SUBMISSION_DATA_PATH}/submission25_0.799.csv")
# submission["prediction"] = test_agg["prediction"].values
submission["prediction"] = test_score_df.applymap(sigmoid).mean(axis=1)
submission.to_csv(f"{SUBMISSION_DATA_PATH}/submission6_0.797055_sigmoid_avg.csv", index=False)

# test_agg = lgbm_dart.get_test_prediction(test_agg)

# submission["prediction"].apply(sigmoid)

test_score_df

#### Take the average of all CV models' predictions

# np.mean(scores_list, axis=0)

test_agg["prediction"] = np.mean(scores_list, axis=0)

def get_metric_columns(df):
    df = df.sort_values(by="prediction", ascending=False)
    df['weight'] = df['target'].apply(lambda x: 20 if x==0 else 1)
    four_pct_cutoff = int(0.04 * df['weight'].sum())
    df['weight_cumsum'] = df['weight'].cumsum()
    df["is_cutoff"] = 0
    df.loc[df['weight_cumsum'] <= four_pct_cutoff, "is_cutoff"] = 1
    df = df.reset_index()
    
    df['random'] = (df['weight'] / df['weight'].sum()).cumsum()
    total_pos = (df['target'] * df['weight']).sum()
    df['cum_pos_found'] = (df['target'] * df['weight']).cumsum()
    df['lorentz'] = df['cum_pos_found'] / total_pos
    df['gini'] = (df['lorentz'] - df['random']) * df['weight']
    return df

test_agg["prediction"].describe()

del scores_list

test_agg.loc[(test_agg["P_2_last"] < 0.4) & (test_agg["prediction"] < 0), "prediction"] += 4

#### Take a look on CV score

with open(f"{EXP_FULL_PATH}/best_scores.json", "r+") as outfile:
    best_scores = json.load(outfile)
np.mean(list(best_scores["validation"].values()))

### Manual Adjustment

test_agg["P_2_last"].describe()

#### Prepare the submission csv, then export out new submission file

submission = pd.read_csv(f"{SUBMISSION_DATA_PATH}/submission25_0.799.csv")
submission["prediction"] = test_agg["prediction"].values
submission.to_csv(f"{SUBMISSION_DATA_PATH}/submission__.csv", index=False)

### Analyze Feature Importance

#### Read old feature importance summary df

old_feature_imp_df = pd.read_csv(f"{EXP_FULL_PATH}/feature_imp_summary.csv")

#### Construct new feature importance summary df

feature_imp_df.loc[feature_imp_df["agg_type"] == ""]

corr_df = pd.read_csv(f"{PROCESSED_DATA_PATH}/raw_column_correlation/train_corr_df.csv")
base_features = corr_df.columns.tolist()

old_feature_imp_df

new_feature_imp_df.to_csv(f"{PROCESSED_DATA_PATH}/feature_imp_summary.csv", index=False)

# validation_indices = joblib.load(f'{EXP_FULL_PATH}/5_fold_val_indices.pkl')

## Feature Importance

# fi_master = read_file(f"{EXP_PATH}/feature_imp_general_summary.csv")

fi_gbdt = lgbm_gbdt2.feature_imp_summary

# fi_gbdt.to_csv(f"{EXP_PATH}/feature_imp_summary5.csv", index=False)

fi_dart = lgbm_dart.feature_imp_df

lgbm_gbdt.feature_imp_summary.to_csv(f"{EXP_PATH}/feature_imp_summary3.csv", index=False)

lgbm_gbdt_shuffled.feature_imp_df.groupby("agg_type")["average_importance"].sum().sort_values().iloc[-30:]

set(fi_gbdt["base_feature"]) - set(fi_dart["base_feature"])

set(fi_dart["base_feature"]) - set(fi_gbdt["base_feature"])

# lgbm_dart.feature_imp_df

imp_df = lgbm_gbdt.feature_imp_df.copy()
imp_summary = lgbm_gbdt.feature_imp_summary.copy()

half_max_imp_df = imp_df.loc[imp_df["average_importance"] > 0.5*imp_df["average_importance"].max()]

top_base_features = imp_summary.loc[imp_summary["base_feature"].isin(half_max_imp_df["base_feature"].tolist())]

plot_heatmap(top_base_features.set_index("base_feature"), annot=True, figsize=(18, 18), fmt=".1f")

lgbm_gbdt.feature_imp_summary.to_csv(f"{EXP_PATH}/feature_imp_summary3.csv", index=False)

noob_features_df = feature_imp_df["average_importance"].nsmallest(int(feature_imp_df.shape[0] * 0.04))
noob_features = noob_features_df.index.tolist()

top_features_df = feature_imp_df["average_importance"].nlargest(int(feature_imp_df.shape[0] * 0.9))
top_features = top_features_df.index.tolist()

pd.Series([f.split("_")[-1] for f in top_features]).value_counts()

pd.Series(top_features, name="feature").to_csv("top_features.csv", index=False)

# used_features = set(feature_imp_df.index.tolist())

mean_features = pad_column_name(MEAN_FEATURES, suffix="_avg")
min_features = pad_column_name(MIN_FEATURES, suffix="_min")
max_features = pad_column_name(MAX_FEATURES, suffix="_max")
last_features = pad_column_name(LAST_FEATURES, suffix="_last")
first_features = pad_column_name(FIRST_FEATURES, suffix="_first")

range_features = pad_column_name(RANGE_FEATURES, suffix="_range")
velocity_features = pad_column_name(VELOCITY_FEATURES, suffix="_velocity")
speed_features = pad_column_name(SPEED_FEATURES, suffix="_speed")

all_features = (
    mean_features + min_features + max_features + last_features + first_features
    + range_features + velocity_features + speed_features
)

len(all_features)

set(top_features) - set(all_features)

top_features

feature_imp_df = feature_imp_df.loc[feature_imp_df["average_importance"] > 0]
feature_imp_subset = feature_imp_df.loc[feature_imp_df["average_importance"] > 200].reset_index()

# feature_imp_subset.loc[feature_imp_subset["feature"].str.contains("last")].sort_values("feature")

t = feature_imp_subset.copy()

t.loc[t["feature"] == "D_128_first"]

# feature_imp_df.index.tolist()

feature_list = feature_imp_df.index.tolist()
len(feature_list)

original_score

simulated_score_list = []
for feature in feature_list[::-1]:
    temp_train_val = train_val.copy()
    temp_train_val[feature] = np.nan
    for i, indices in enumerate(validation_indices):
        temp_model = models[i]
        temp_val = temp_train_val.loc[indices, temp_model.feature_name_]
        temp_train_val.loc[indices, "score"] = temp_model.predict_proba(temp_val, raw_score=True)
    simulated_score = amex_metric(temp_train_val["target"], temp_train_val["score"])
    simulated_score_list.append(simulated_score)
    if simulated_score >= original_score:
        sign = "+"
    else:
        sign = ""
    if abs(simulated_score - original_score) >= 1e-4:
        print(feature, f" {sign}{(simulated_score - original_score):.4f}")
    print(f"{feature} done")
    del temp_train_val
    gc.collect()

test = read_file(f"{RAW_TEST_PICKLE_PATH}/test_data.pkl")

test.loc[test["row_number"] > 12]

test_agg = process_test_data(test)

test_agg.head()

del test

for i, model in enumerate(models.values()):
    model_input = test_agg.loc[:, model.feature_name_]
    test_agg.loc[:, f"score{i}"] = model.predict_proba(model_input, raw_score=True)

test_agg.head()

new_result_sub = pd.read_csv(f"{SUBMISSION_DATA_PATH}/submission19_0.796.csv")

old_result_sub = pd.read_csv(f"{SUBMISSION_DATA_PATH}/submission17_0.796.csv")

plt.scatter(old_result_sub["prediction"], new_result_sub["prediction"])

np.corrcoef(old_result_sub["prediction"], new_result_sub["prediction"])

merged_result = new_result_sub.merge(old_result_sub, on="customer_ID", how="left")

merged_result["x_positive"] = (merged_result["prediction_x"] >= 0)

merged_result["y_positive"] = (merged_result["prediction_y"] >= 0)

neg_neg_indices = merged_result.loc[(~merged_result["x_positive"]) & (~merged_result["y_positive"])].index

merged_result.loc[neg_neg_indices, "prediction"] = merged_result.loc[neg_neg_indices, :].min(axis=1)

pos_pos_indices = merged_result.loc[(merged_result["x_positive"]) & (merged_result["y_positive"])].index

merged_result.loc[pos_pos_indices, "prediction"] = merged_result.loc[pos_pos_indices, :].max(axis=1)

merged_result["prediction"].isnull().sum()

merged_result.loc[merged_result["prediction"].isnull(), "prediction"] = \
(merged_result.loc[merged_result["prediction"].isnull(), "prediction_x"] + \
merged_result.loc[merged_result["prediction"].isnull(), "prediction_y"]) / 2

merged_result = merged_result.drop(columns=["prediction_x", "prediction_y", "x_positive", "y_positive"])

merged_result.to_csv(f"{SUBMISSION_DATA_PATH}/submission23.csv", index=False)

new_result_sub["prediction"] = np.max([new_result_sub["prediction"].values, old_result_sub["prediction"].values], axis=0)

new_result_sub["prediction"] += old_result_sub["prediction"]

new_result_sub["prediction"] /= 2

new_result_sub.to_csv(f"{SUBMISSION_DATA_PATH}/submission22.csv", index=False)

result_sub = test_agg.loc[:, [c for c in test_agg.columns if c.startswith("score")]].mean(axis=1).reset_index()

result_sub.columns = ["customer_ID", "prediction"]

result_sub.to_csv("submission20.csv", index=False)

single_record_index = test_agg.loc[test_agg["num_records"] == 1].index.tolist()

final_df13 = pd.read_csv(f"{SUBMISSION_DATA_PATH}/submission13_0.796.csv")
final_df13["prediction_13"] = sigmoid(final_df13["prediction"])
final_df13 = final_df13.drop(columns="prediction")

final_df16 = pd.read_csv(f"{SUBMISSION_DATA_PATH}/submission16_0.782.csv")
final_df16["prediction_16"] = sigmoid(final_df16["prediction"])
final_df16 = final_df16.drop(columns="prediction")

mega_df = final_df13.merge(final_df16, on="customer_ID", how="left")

mega_df.loc[:, "prediction"] = (
    mega_df.loc[:, "prediction_16"] * 0.1 \
    + mega_df.loc[:, "prediction_13"] * 0.9
)

mega_df.loc[single_record_index, "prediction"] = (
    mega_df.loc[single_record_index, "prediction_16"] * 0.8 \
    + mega_df.loc[single_record_index, "prediction_13"] * 0.2
)

mega_df = mega_df.drop(columns=["prediction_13", "prediction_16"])

mega_df["prediction"] = 0.55 * mega_df["prediction_x"] + 0.45 * mega_df["prediction_y"]
mega_df = mega_df.drop(columns=["prediction_x", "prediction_y"])

mega_df.to_csv(f"{SUBMISSION_DATA_PATH}/submission18.csv", index=False)
