==> notebooks/5z.local_evaluation.ipynb <==
import gc
import joblib
import matplotlib.pyplot as plt
import pandas as pd
import numpy as np
import os
import sys
sys.path.append("../")
import time
import warnings
warnings.simplefilter("ignore")
from lightgbm import LGBMClassifier
from sklearn.preprocessing import MinMaxScaler
from sklearn.model_selection import train_test_split
from sklearn.metrics import roc_auc_score, fbeta_score, make_scorer
from tqdm import tqdm

from utils.common import (
    sigmoid, reverse_sigmoid,
    pad_column_name
)
from utils.constants import *
from utils.eval_helpers import (
    plot_roc_curves, plot_feature_importance, 
    amex_metric, get_final_metric_df, amex_metric_np, lgb_amex_metric
)
from utils.eda_helpers import (
    plot_missing_proportion_barchart, 
    get_cols
)
from utils.extraction_helpers import read_file
from utils.feature_group import (
    CATEGORY_COLUMNS, CONTINUOUS_COLUMNS, NON_FEATURE_COLUMNS
)

%load_ext autoreload
%autoreload

### Read Data

train_score1 = read_file(f"{EXP_PATH}/1.lgbm_dart_923/train_scores_df.csv")
train_score2 = read_file(f"{EXP_PATH}/2.lgbm_dart_1020/train_scores_df.csv")
train_score3 = read_file(f"{EXP_PATH}/3.lgbm_dart_5197/train_scores_df.csv")
train_score4 = read_file(f"{EXP_PATH}/4.lgbm_dart_6666/train_scores_df.csv")
train_score4 = read_file(f"{EXP_PATH}/5.lgbm_dart_923_half_fix/train_scores_df.csv")

sum(train_score3["target"].values != train_score4["target"].values)

t1 = train_score1["cv_score"].values
t2 = train_score2["cv_score"].values
t3 = train_score3["cv_score"].values
t4 = train_score4["cv_score"].values
t5 = train_score4["cv_score"].values

## 1ST - 0.799 (medium)
amex_metric(train_score1["target"], t1)

## 3RD - 0.800 (low)
amex_metric(train_score1["target"], t2)

## 5TH - 0.799 (low)
amex_metric(train_score1["target"], t3)

### 6TH - 0.800 (very low)
amex_metric(train_score1["target"], t4)

### 7TH - 0.798 (high)
amex_metric(train_score1["target"], t5)

from collections import defaultdict
from itertools import product

m = MinMaxScaler()

t1_ = m.fit_transform(t1.reshape(-1, 1))[:, 0]
t2_ = m.fit_transform(t2.reshape(-1, 1))[:, 0]
t3_ = m.fit_transform(t3.reshape(-1, 1))[:, 0]
t4_ = m.fit_transform(t4.reshape(-1, 1))[:, 0]
t5_ = m.fit_transform(t5.reshape(-1, 1))[:, 0]

sign_series = np.where(sum([(t1 - 20) > 0, (t2 - 20) > 0, t3 > 0, (t4 > 0)]) >= 2, 1, -1)

abs_series = t1 * t2 * t3 * t4

blend = abs_series * sign_series

amex_metric(train_score1["target"], blend)

x = np.arange(0.1, 0.6, 0.1)

d = defaultdict(list)
for i, j, k, l, m in product(x, x, x, x, x):
    d["i"].append(i)
    d["j"].append(j)
    d["k"].append(k)
    d["l"].append(l)
    d["m"].append(m)
    # blend = i * sigmoid(t1) + j * sigmoid(t2) + k * sigmoid(t3) + l * sigmoid(t4)
    blend = i * t1_ + j * t2_ + k * t3_ + l * t4_ + m * t5_
    overall, gini, top4pct = amex_metric(train_score1["target"], blend)
    d["score"].append(overall)
    d["gini"].append(gini)
    d["top4pct"].append(top4pct)

result = pd.DataFrame(d)
# result.nsmallest(50, "score")

result.nlargest(10, "score")

result.nlargest(10, "score")

dd = defaultdict(list)
for i in np.arange(0.46, 0.54, 0.01):
    for j in np.arange(0.22, 0.28, 0.01):
        for k in np.arange(0.22, 0.28, 0.01):
            dd["i"].append(i)
            dd["j"].append(j)
            dd["k"].append(k)
            blend = i*t1 + j*t2 + k*t3
            overall, gini, top4pct = amex_metric(train_score1["target"], blend)
            dd["score"].append(overall)
            dd["gini"].append(gini)
            dd["top4pct"].append(top4pct)

result = pd.DataFrame(dd)
result.nlargest(5, "score")

sub = pd.read_csv(f"{SUBMISSION_DATA_PATH}/empty.csv")
sub["prediction"] = pd.Series(
    (0.9 * t1) + 
    (0.7 * t2) + 
    (0.4 * t3) + 
    (0.6 * t4)
)
sub.to_csv(f"{SUBMISSION_DATA_PATH}/blended_train_scores.csv", index=False)

### BLEND TEST

test_score1 = read_file(f"{EXP_PATH}/1.lgbm_dart_923/scores_df.csv")
test_score2 = read_file(f"{EXP_PATH}/2.lgbm_dart_1020/scores_df.csv")
test_score3 = read_file(f"{EXP_PATH}/3.lgbm_dart_5197/scores_df.csv")
test_score4 = read_file(f"{EXP_PATH}/4.lgbm_dart_6666/scores_df.csv")
test_score5 = read_file(f"{EXP_PATH}/5.lgbm_dart_923_half_fix/scores_df.csv")

s = read_file(f"{SUBMISSION_DATA_PATH}/submission3_lgbm_dart_full_fix_1020.csv")["prediction"]

s2 = test_score2.mean(axis=1)

m = MinMaxScaler()

test_score2_ = m.fit_transform(test_score2)

np.mean(test_score2_, axis=1)

sub = pd.read_csv(f"{SUBMISSION_DATA_PATH}/empty.csv")
sub["prediction"] = np.mean(test_score2_, axis=1)
sub.to_csv(f"{SUBMISSION_DATA_PATH}/submission_9TH_seed_1020_average_fold_after_minmaxscaler.csv", index=False)

sub = pd.read_csv(f"{SUBMISSION_DATA_PATH}/empty.csv")
sub["prediction"] = (
    (0.4 * m.fit_transform(test_score1).mean(axis=1)) + 
    (0.1 * m.fit_transform(test_score2).mean(axis=1)) + 
    (0.2 * m.fit_transform(test_score3).mean(axis=1)) + 
    (0.1 * m.fit_transform(test_score4).mean(axis=1)) +
    (0.1 * m.fit_transform(test_score5).mean(axis=1))
) / 0.9
sub.to_csv(f"{SUBMISSION_DATA_PATH}/submission_10TH_blend_5_own_seeds_after_min_max_scaled.csv", index=False)

sub["prediction"].describe()
