==> notebooks/6c.generate_difficulty_group.ipynb <==
import gc
import joblib
import matplotlib.pyplot as plt
import pandas as pd
import numpy as np
import os
import seaborn as sns
import sys
sys.path.append("../")
import time
import warnings
warnings.simplefilter("ignore")
from lightgbm import LGBMClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.neighbors import KNeighborsClassifier
from sklearn.preprocessing import MinMaxScaler, StandardScaler
from sklearn.model_selection import train_test_split, StratifiedKFold
from sklearn.metrics import roc_auc_score, fbeta_score, make_scorer
from tqdm import tqdm

from utils.common import (
    sigmoid, reverse_sigmoid,
    pad_column_name
)
from utils.constants import *
from utils.eval_helpers import (
    plot_roc_curves, plot_feature_importance, get_final_metric_df,
    amex_metric, get_final_metric_df, amex_metric_np, lgb_amex_metric, 
    plot_score_distribution, train_logistic_regression
)
from utils.eda_helpers import (
    plot_missing_proportion_barchart, 
    get_cols, plot_heatmap
)
from utils.extraction_helpers import read_file
from utils.feature_group import (
    CATEGORY_COLUMNS, CONTINUOUS_COLUMNS, NON_FEATURE_COLUMNS
)

%load_ext autoreload
%autoreload

ID_COLUMNS = ["customer_ID", "target"]

### Read Data

labels = read_file(f"{RAW_DATA_PATH}/train_labels.csv")
target = labels["target"]
labels.head(2)

num_statements_df = read_file(f"{RAW_DATA_PATH}/train_num_statements.csv")
num_statements_df.head(2)

master_train_score_df = read_file(f"{EXP_PATH}/master_result/master_train_scores.csv")
master_train_score_df = master_train_score_df.loc[:, ID_COLUMNS + get_cols(master_train_score_df, "oof")]

train_score_c1 = read_file(f"{EXP_PATH}/chris/0.792648_custom_emb_local_best_no_sigmoid_skf_oof.csv")
train_score_c1 = train_score_c1.rename(columns={"oof_prediction": "chris_exp1_oof_score"})

master_train_score_df = master_train_score_df.merge(
    train_score_c1[ID_COLUMNS + ["chris_exp1_oof_score"]], on=ID_COLUMNS, how='left'
)

oof_columns = get_cols(master_train_score_df, "oof")

for col in oof_columns:
    metric = amex_metric(master_train_score_df["target"], master_train_score_df[col])
    print(f"{col.split('_')[0]}: OOF amex metric - {metric[0]:.5f}")

normal_df = master_train_score_df.loc[master_train_score_df["target"] == 0].reset_index(drop=True)
default_df = master_train_score_df.loc[master_train_score_df["target"] == 1].reset_index(drop=True)

normal_df.shape, default_df.shape

### Get Normal Group

- Group = 0 means easiest & Group = 4 means the hardest to predict this user as a NORMAL user

normal_df.columns

selected_columns = [f"exp{i}_oof_score" for i in [1, 2, 4, 9]]

normal_df["train_oof"] = normal_df[selected_columns].mean(axis=1)

normal_df["group"] = pd.qcut(normal_df["train_oof"], q=5).cat.codes

normal_df = normal_df[["customer_ID", "target", "group"]]

normal_df.to_csv(f"{EXP_PATH}/master_result/normal_predict_group.csv", index=False)

### Get Default Group

- Group = 0 means easiest & Group = 4 means the hardest to predict this user as a DEFAULT user

default_df.columns

selected_columns = [f"exp{i}_oof_score" for i in [1, 2, 4, 9]]

default_df["train_oof"] = default_df[selected_columns].mean(axis=1)

default_df["group"] = 4 - pd.qcut(default_df["train_oof"], q=5).cat.codes

default_df = default_df[["customer_ID", "target", "group"]]

default_df["group"].value_counts()

default_df.to_csv(f"{EXP_PATH}/master_result/default_predict_group.csv", index=False)
