==> notebooks/7a.gambling4.ipynb <==
import gc
import joblib
import matplotlib.pyplot as plt
import pandas as pd
import numpy as np
import os
import sys
sys.path.append("../")
import time
import warnings
warnings.simplefilter("ignore")
from lightgbm import LGBMClassifier
from sklearn.preprocessing import MinMaxScaler
from sklearn.model_selection import train_test_split
from tqdm import tqdm

from utils.common import sigmoid, reverse_sigmoid
from utils.constants import *
from utils.extraction_helpers import read_file

%load_ext autoreload
%autoreload

### Gambling13

gambling6 = read_file(f"{SUBMISSION_DATA_PATH}/gambling6_0.801.csv")
# waikey["prediction2"] = read_file(f"{EXP_PATH}/16.lgbm_dart_6666_new/scores_df.csv").mean(axis=1)

gambling6["prediction"] = gambling6["prediction"].apply(reverse_sigmoid)

new = read_file(f"{SUBMISSION_DATA_PATH}/first_real_blending_of_dart_and_rnn.csv")

new["prediction"] = new["prediction"].apply(reverse_sigmoid)

pd.concat([gambling6["prediction"], new["prediction"]], axis=1).describe()

# fig, ax = plt.subplots(figsize=(12, 12))
# # plt.plot(np.arange(0, 1, 0.01), np.arange(0, 1, 0.01), color="orange", linestyle="-")
# # plt.xlabel("WK Blend")
# # plt.ylabel("Chris best NN")
# plt.scatter(gambling6["prediction"], new["prediction"])
# # plt.scatter(s1, s3)
# plt.show()

w = 0.85 * (gambling6["prediction"] - 0.33).apply(sigmoid) + 0.15 * new["prediction"].apply(sigmoid)

sub = pd.read_csv(f"{SUBMISSION_DATA_PATH}/empty.csv")
sub["prediction"] = w
sub.to_csv(f"{SUBMISSION_DATA_PATH}/0.85_gambling6_0.15_blend.csv", index=False)

### Gambling15

chris  = read_file(f"{SUBMISSION_DATA_PATH}/sorted_chris.csv")
chris_best  = read_file(f"{SUBMISSION_DATA_PATH}/chris_best.csv")
waikey = read_file(f"{SUBMISSION_DATA_PATH}/wk_seed_6666.csv")
waikey2 = read_file(f"{SUBMISSION_DATA_PATH}/wk_seed_1020.csv")
ensemble = read_file(f"{SUBMISSION_DATA_PATH}/others_ensemble.csv")
exp_ensemble = read_file(f"{SUBMISSION_DATA_PATH}/others_exponential_ensemble.csv")
blend_trees = read_file(f"{SUBMISSION_DATA_PATH}/others_blend_boosted_tree.csv")

chris["prediction"] = 0.5 * chris["prediction"] + 0.5 * chris_best["prediction"]

waikey["sigmoid_prediction"] = (0.5 * waikey["prediction"] + 0.5 * waikey2["prediction"]).apply(sigmoid)

pre_blend_df = pd.DataFrame(
    dict(
        wk=waikey["sigmoid_prediction"],
        chris=chris["prediction"],
        ensemble=ensemble["prediction"],
        exp_ensemble=exp_ensemble["prediction"],
        blend_trees=blend_trees["prediction"]
    )
)

pre_blend_df.describe()

mscaler = MinMaxScaler()
pp = mscaler.fit_transform(pre_blend_df)

pre_blend_df = pd.DataFrame(pp, columns=pre_blend_df.columns)

pre_blend_df.describe()

pre_blend_df.corr()

pre_blend_df["final1"] = (
    0.57 * pre_blend_df["wk"] + 0.17 * pre_blend_df["chris"] + 0.26 * pre_blend_df["ensemble"]
)

sub = pd.read_csv(f"{SUBMISSION_DATA_PATH}/empty.csv")
sub["prediction"] = pre_blend_df["final1"]
sub.to_csv(f"{SUBMISSION_DATA_PATH}/gambling16.csv", index=False)
